+++ packages/terra-icon/examples/test-examples/IconDefault.example.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import IconAdd from '../../../lib/icon/IconAdd';
+import IconAdd from '../../lib/icon/IconAdd';
 
 const IconDefault = () => (
   <div>